+++ quint/test/parsing/quintParserFrontend.test.ts
@@ -8,6 +8,7 @@ import {
   parsePhase1fromText,
   parsePhase2sourceResolution,
   parsePhase3importAndNameResolution,
+  parsePhase4toposort,
 } from '../../src/parsing/quintParserFrontend'
 import { lf } from 'eol'
 import { right } from '@sweet-monads/either'
@@ -30,6 +31,12 @@ function readJson(name: string): any {
 
 // read the Quint file and the expected JSON, parse and compare the results
 function parseAndCompare(artifact: string): void {
+  // read the expected result as JSON
+  const expected = readJson(artifact)
+  // We're not interested in testing the contens of the table here
+  delete expected.table
+  let outputToCompare
+
   // read the input from the data directory and parse it
   const gen = newIdGenerator()
   const basepath = resolve(__dirname, '../../testFixture')
@@ -39,18 +46,13 @@ function parseAndCompare(artifact: string): void {
     return path.replace(basepath, 'mocked_path/testFixture')
   })
   const mainPath = resolver.lookupPath(basepath, `${artifact}.qnt`)
-  const phase1Result = parsePhase1fromText(gen, readQuint(artifact), mainPath.toSourceName()).chain(res =>
+  const phase2Result = parsePhase1fromText(gen, readQuint(artifact), mainPath.toSourceName()).chain(res =>
     parsePhase2sourceResolution(gen, resolver, mainPath, res)
   )
-  // read the expected result as JSON
-  const expected = readJson(artifact)
-  // We're not interested in testing the contens of the table here
-  delete expected.table
-  let outputToCompare
 
-  if (phase1Result.isLeft()) {
-    // An error occurred at phase 1, check if it is the expected result
-    phase1Result.mapLeft(
+  if (phase2Result.isLeft()) {
+    // An error occurred at phase 2, check if it is the expected result
+    phase2Result.mapLeft(
       err =>
         (outputToCompare = {
           stage: 'parsing',
@@ -58,10 +60,10 @@ function parseAndCompare(artifact: string): void {
           warnings: [],
         })
     )
-  } else if (phase1Result.isRight()) {
-    const { modules, sourceMap } = phase1Result.value
-    const expectedIds = modules.flatMap(m => collectIds(m)).sort()
-    // Phase 1 succeded, check that the source map is correct
+  } else if (phase2Result.isRight()) {
+    const { modules: modules2, sourceMap } = phase2Result.value
+    const expectedIds = modules2.flatMap(m => collectIds(m)).sort()
+    // Phase 1-2 succededed, check that the source map is correct
     assert.sameDeepMembers(expectedIds, [...sourceMap.keys()].sort(), 'expected source map to contain all ids')
 
     const expectedSourceMap = readJson(`${artifact}.map`)
@@ -69,11 +71,13 @@ function parseAndCompare(artifact: string): void {
 
     assert.deepEqual(sourceMapResult, expectedSourceMap, 'expected source maps to be equal')
 
-    const phase3Result = parsePhase3importAndNameResolution(phase1Result.value)
+    const phase4Result = parsePhase3importAndNameResolution(phase2Result.value).chain(phase3Data =>
+      parsePhase4toposort(phase3Data)
+    )
 
-    if (phase3Result.isLeft()) {
-      // An error occurred at phase 2, check if it is the expected result
-      phase3Result.mapLeft(
+    if (phase4Result.isLeft()) {
+      // An error occurred at phases 3-4, check if it is the expected result
+      phase4Result.mapLeft(
         err =>
           (outputToCompare = {
             stage: 'parsing',
@@ -82,8 +86,9 @@ function parseAndCompare(artifact: string): void {
           })
       )
     } else {
-      // Both phases succeeded, check that the module is correclty outputed
-      outputToCompare = { stage: 'parsing', warnings: [], modules: modules }
+      // All phases succeeded, check that the module is correclty output
+      const modules4 = phase4Result.value.modules
+      outputToCompare = { stage: 'parsing', warnings: [], modules: modules4 }
     }
   }
 
@@ -118,6 +123,10 @@ describe('parsing', () => {
 
   it('parses SuperSpec correctly', () => {
     parseAndCompare('SuperSpec')
+  })
+
+  it('parses out of order definitions', () => {
+    parseAndCompare('_0099unorderedDefs')
   })
 })
 
@@ -211,5 +220,13 @@ describe('parse errors', () => {
 
   it('error on overriding values that are not constants', () => {
     parseAndCompare('_1016nonConstOverride')
+  })
+
+  it('error on cyclic definitions', () => {
+    parseAndCompare('_0100cyclicDefs')
+  })
+
+  it('error on accidental recursion', () => {
+    parseAndCompare('_0101noRecursion')
   })
 })